requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
pyyaml>=5.4.1
pytest>=6.2.5
logging>=0.5.1.2 